+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Improvements for link preview
 - 'Copy' commands now display browser notification
 - Added new option to show hint on hovering action when it's icon-only (default: on)
+- Improved 'long left click to open' option
 
 #1.2.1
 - Improved 'long left click to open menu' functionality